feature: technician module

+++ src/database/entities/base.entity.ts
@@ -2,6 +2,6 @@ import { Timestamp } from '@src/database/entities/timestamp.entity';
 import { PrimaryGeneratedColumn } from 'typeorm';
 
 export abstract class BaseEntity extends Timestamp {
-  @PrimaryGeneratedColumn({ type: 'bigint', unsigned: true })
+  @PrimaryGeneratedColumn({ type: 'int', unsigned: true })
   id: number;
 }
